# doc/电商平台.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "电商平台",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/ordersItem/findByOrdersId": {
      "get": {
        "summary": "通过订单ID查找订单商品",
        "x-apifox-folder": "订单管理",
        "x-apifox-status": "deprecated",
        "deprecated": true,
        "description": "",
        "tags": [
          "订单管理"
        ],
        "parameters": [
          {
            "name": "orders_id",
            "in": "query",
            "description": "订单ID",
            "required": false,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "quantity": {
                            "type": "integer"
                          },
                          "product_id": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "quantity",
                          "product_id"
                        ],
                        "x-apifox-orders": [
                          "quantity",
                          "product_id"
                        ],
                        "x-apifox-ignore-properties": []
                      },
                      "nullable": true
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "success": true,
                      "message": null,
                      "data": [
                        {
                          "id": 302,
                          "user": {
                            "id": 52,
                            "name": "czy",
                            "password": "czy",
                            "gender": "男",
                            "email": "777Test@qq.com",
                            "phone": "12312312333",
                            "address": "北京交通大学",
                            "token": ""
                          },
                          "isPayed": false,
                          "isDelivered": false,
                          "isFinished": false
                        }
                      ],
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2902459/apis/api-90714956-run"
      }
    },
...
    "/ordersItem/findDetailsByOrderId": {
      "get": {
        "summary": "通过订单ID查找订单商品详情",
        "x-apifox-folder": "订单管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据当前订单ID查找商品并展示商品详情，对订单商品不存在有异常抛出判断",
        "tags": [
          "订单管理"
        ],
        "parameters": [
          {
            "name": "orders_id",
            "in": "query",
            "description": "订单ID",
            "required": false,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "price": {
                            "type": "integer"
                          },
                          "isList": {
                            "type": "boolean"
                          },
                          "shop": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "shopName": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "shopName"
                            ],
                            "x-apifox-orders": [
                              "id",
                              "shopName"
                            ],
                            "x-apifox-ignore-properties": []
                          },
                          "category": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "categoryName": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "categoryName"
                            ],
                            "x-apifox-orders": [
                              "id",
                              "categoryName"
                            ],
                            "x-apifox-ignore-properties": []
                          }
                        },
                        "x-apifox-orders": [
                          "id",
                          "name",
                          "description",
                          "price",
                          "isList",
                          "shop",
                          "category"
                        ],
                        "x-apifox-ignore-properties": []
                      },
                      "nullable": true
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "success": true,
                      "message": null,
                      "data": [
                        {
                          "product": {
                            "id": 1,
                            "name": "test1",
                            "description": "test2",
                            "price": 234,
                            "isList": true,
                            "rate": 0,
                            "shop": {
                              "id": 1,
                              "shopName": "test1"
                            },
                            "category": {
                              "id": 1,
                              "categoryName": "test1"
                            }
                          },
                          "quantity": 5
                        }
                      ],
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2902459/apis/api-90722285-run"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "localhost:8080",
      "description": "测试环境"
    }
  ]
}